+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.1] - 2014-05-31
 ### Added
-- Added JWT based security.
+- Added JWT based security and user management.
+  - Added basic user registration endpoint - save to db.
+  - Added basic user authentication endpoint to generate JWT token once user is validated.
 
 [Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.0...HEAD
 [0.0.1]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.0.1